+++ utiles/include/Json.h
@@ -30,12 +30,12 @@ public:
     virtual void reset();
 
     virtual void agregarAtributoValor(std::string clave, std::string valor);
-    virtual void agregarAtributoValor(std::string clave, unsigned long long int valor);
-    virtual void agregarAtributoValor(std::string clave, unsigned int valor);
+    virtual void agregarAtributoValor(std::string clave, uint64_t valor);
+    virtual void agregarAtributoValor(std::string clave, uint32_t valor);
     virtual void agregarAtributoValor(std::string clave, float valor);
     virtual void agregarAtributoValor(std::string clave, bool valor);
 
-    virtual void agregarAtributoArray(std::string clave, std::vector<unsigned long long int> array_valores);
+    virtual void agregarAtributoArray(std::string clave, std::vector<uint64_t> array_valores);
     virtual void agregarAtributoArray(std::string clave, std::vector<std::string> array_valores);
     virtual void agregarAtributoArray(std::string clave, std::vector<Json*> array_valores);
 
@@ -46,14 +46,14 @@ public:
     // GETTERS
 
     virtual std::string getAtributoValorString(std::string clave);
-    virtual unsigned long long int getAtributoValorUint(std::string clave);
+    virtual uint64_t getAtributoValorUint(std::string clave);
     virtual float getAtributoValorFloat(std::string clave);
     virtual bool getAtributoValorBool(std::string clave);
 
     // Devuelve una copia. Cuando se termine de usar, se tiene que destruir (usando 'delete').
     virtual Json* getAtributoValorJson(std::string clave);
 
-    virtual std::vector<unsigned long long int> getAtributoArrayUint(std::string clave);
+    virtual std::vector<uint64_t> getAtributoArrayUint(std::string clave);
     virtual std::vector<std::string> getAtributoArrayString(std::string clave);
 
     // Devuelve una copia de los valores. Cuando se terminen de usar, se tienen que destruir (usando 'delete').
@@ -69,7 +69,7 @@ public:
 
     virtual void setValor(rapidjson::Value* valor);
 
-    virtual void setCantidadMaximaDecimales(unsigned int cantidad_maxima_decimales);
+    virtual void setCantidadMaximaDecimales(uint32_t cantidad_maxima_decimales);
 
     // CONSULTA
 
@@ -85,7 +85,7 @@ private:
 
     std::vector<Json*> copia_atributos_json;
 
-    unsigned int cantidad_maxima_decimales;
+    uint32_t cantidad_maxima_decimales;
 };
 
 };
