+++ utiles/source/GestorIDs.cpp
@@ -4,7 +4,7 @@ using namespace herramientas::utiles;
 
 // CONSTRUCTORES
 
-GestorIDs::GestorIDs(unsigned long long int id_mas_alto) : id_mas_alto(id_mas_alto)
+GestorIDs::GestorIDs(uint64_t id_mas_alto) : id_mas_alto(id_mas_alto)
 {
 }
 
@@ -14,7 +14,7 @@ GestorIDs::~GestorIDs()
 
 // SETTERS
 
-void GestorIDs::setIdActual(unsigned long long int id_mas_alto)
+void GestorIDs::setIdActual(uint64_t id_mas_alto)
 {
 	GestorIDs::id_mas_alto = id_mas_alto;
 }
@@ -28,7 +28,7 @@ ID * GestorIDs::nuevoID()
 	return id_nuevo;
 }
 
-unsigned long long int GestorIDs::getIdActual()
+uint64_t GestorIDs::getIdActual()
 {
 	return id_mas_alto;
 }
