+++ utiles/source/IHashable.cpp
@@ -3,7 +3,7 @@
 using namespace herramientas::utiles;
 
 std::hash<std::string> IHashable::hasher_string;
-std::hash<unsigned int> IHashable::hasher_uint;
+std::hash<uint32_t> IHashable::hasher_uint;
 
 IHashable::IHashable()
 {
@@ -13,12 +13,12 @@ IHashable::~IHashable()
 {
 }
 
-unsigned long long int IHashable::hashear(std::string string_a_hashear)
+uint64_t IHashable::hashear(std::string string_a_hashear)
 {
     return hasher_string(string_a_hashear);
 }
 
-unsigned long long int IHashable::hashear(unsigned int uint_a_hashear)
+uint64_t IHashable::hashear(uint32_t uint_a_hashear)
 {
     return hasher_uint(uint_a_hashear);
 }
